Layout and not-found page refactor, added css parameters and preview-contacto image upscaled.
